Add Helius DAS API support in DN and refactor
