+++ packages/discovery-provider/default_config.ini
@@ -48,6 +48,9 @@ rewards_manager_min_slot = 0
 anchor_data_program_id = 6znDH9AxEi9RSeDR7bt9PVYRUS4XxZLKhni96io9Aykb
 anchor_admin_storage_public_key = 9Urkpt297u2BmLRpNrwsudDjK6jjcWxTaDZtyS2NRuqX
 
+[nft]
+helius_das_api_url = https://sol-collectibles.audius.co
+
 [redis]
 url = redis://localhost:5379/0
